+++ copernican/lib/likelihoods/cmb/__init__.py
@@ -3,9 +3,11 @@
 from .cmb import (
     CMBLike,
     compute_cmb_spectrum,
+    compute_cmb_spectrum_batch,
     compute_cmb_spectrum_cached,
     compute_cmb_spectrum_from_contract,
 )
+from .native_batch import NativeCMBBatchResult
 from .native_errors import (
     NativeCMBError,
     NativeConstraintViolationError,
@@ -21,6 +23,7 @@ from .native_errors import (
 
 __all__ = [
     "CMBLike",
+    "NativeCMBBatchResult",
     "NativeCMBError",
     "NativeConstraintViolationError",
     "NativeContractError",
@@ -32,6 +35,7 @@ __all__ = [
     "NativePerformanceBudgetError",
     "NativeUnsupportedCapabilityError",
     "compute_cmb_spectrum",
+    "compute_cmb_spectrum_batch",
     "compute_cmb_spectrum_cached",
     "compute_cmb_spectrum_from_contract",
 ]